[template] refactor(settings): renaming configs to settings
