+++ template/src/models/__init__.py
@@ -1,3 +1,3 @@
-from src.models.configs.config_loader import ConfigLoader
+from src.models.settings.config_loader import ConfigLoader
 
 __all__ = ["ConfigLoader"]
