+++ template/src/models/settings/collector_configs.py
